refactor: argTypes
