+++ src/crypto-providers/trezorCryptoProvider.ts
@@ -48,7 +48,7 @@ import { Errors } from '../errors'
 import { partition } from '../util'
 import { KesVKey, OpCertIssueCounter, SignedOpCertCborHex } from '../opCert/opCert'
 import { parseBIP32Path } from '../command-parser/parsers'
-import { ParsedShowAddressArguments } from '../types'
+import { ParsedShowAddressArguments } from '../argTypes'
 
 const { bech32 } = require('cardano-crypto.js')
